Organize notifications, namespacing Also drops `MemberRequestDenied` for now
