+++ app/Http/Controllers/TicketController.php
@@ -6,9 +6,9 @@ use App\Http\Requests\CreateTicket;
 use App\Models\Ticket;
 use App\Models\TicketType;
 use App\Models\User;
-use App\Notifications\NotifyCallerTicketUpdated;
-use App\Notifications\NotifyNewTicketOwner;
-use App\Notifications\TicketReaction;
+use App\Notifications\Channel\NotifyCallerTicketUpdated;
+use App\Notifications\DM\NotifyNewTicketOwner;
+use App\Notifications\React\TicketReaction;
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Response;
