+++ app/Notifications/Channel/NotifyAdminTicketUpdated.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Notifications;
+namespace App\Notifications\Channel;
 
 use App\Channels\BotChannel;
 use App\Channels\Messages\BotDMMessage;
