+++ app/Notifications/Channel/NotifyDivisionMemberNameChanged.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Notifications;
+namespace App\Notifications\Channel;
 
 use App\Channels\BotChannel;
 use App\Channels\Messages\BotChannelMessage;
@@ -10,7 +10,7 @@ use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Notification;
 
-class MemberNameChanged extends Notification implements ShouldQueue
+class NotifyDivisionMemberNameChanged extends Notification implements ShouldQueue
 {
     use Queueable, RetryableNotification;
