+++ app/Notifications/Channel/NotifyDivisionMemberRemoved.php
@@ -1,33 +1,38 @@
 <?php
 
-namespace App\Notifications;
+namespace App\Notifications\Channel;
 
 use App\Channels\BotChannel;
 use App\Channels\Messages\BotChannelMessage;
+use App\Models\Division;
 use App\Models\Member;
+use App\Models\Squad;
 use App\Models\User;
 use App\Traits\DivisionSettableNotification;
 use App\Traits\RetryableNotification;
-use Exception;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Notification;
 
-class NewExternalRecruit extends Notification implements ShouldQueue
+class NotifyDivisionMemberRemoved extends Notification implements ShouldQueue
 {
     use DivisionSettableNotification, Queueable, RetryableNotification;
 
-    private string $alertSetting = 'chat_alerts.member_created';
+    private string $alertSetting = 'chat_alerts.member_removed';
 
     /**
      * Create a new notification instance.
      */
-    public function __construct(private readonly Member $member, private readonly User $recruiter) {}
+    public function __construct(
+        private readonly Member $member,
+        private readonly User $remover,
+        private readonly string $removalReason,
+        private readonly ?Squad $squad = null
+    ) {}
 
     /**
      * Get the notification's delivery channels.
      *
-     * @param  mixed  $notifiable
      * @return array
      */
     public function via()
@@ -39,47 +44,49 @@ class NewExternalRecruit extends Notification implements ShouldQueue
      * @param  mixed  $notifiable
      * @return array
      *
-     * @throws Exception
+     * @throws \Exception
      */
-    public function toBot($notifiable)
+    public function toBot(Division $notifiable)
     {
-        $recruiter = $this->recruiter;
-        $handle = $this->member->handles->filter(fn ($handle) => $handle->id === $this->member->division->handle_id)->first();
+        $remover = $this->remover;
+        $handle = $this->member->handles->filter(fn ($handle) => $handle->id === $notifiable->handle_id)->first();
 
         return (new BotChannelMessage($notifiable))
             ->title($notifiable->name . ' Division')
             ->target($notifiable->settings()->get($this->alertSetting))
-            ->thumbnail($notifiable->getLogoPath())
-            ->fields([
+            ->thumbnail($notifiable->getLogoPath())->fields([
                 [
-                    'name' => ':crossed_swords: New Member Recruited (External)',
+                    'name' => 'Member Removed',
                     'value' => sprintf(
-                        '%s from %s just recruited [%s](%s) into the %s Division!',
-                        $recruiter->name,
-                        $recruiter->member->division->name,
+                        ':door: [%s](%s) [%d] was removed from %s by %s.',
                         $this->member->name,
                         route('member', $this->member->getUrlParams()),
-                        $notifiable->name
+                        $this->member->clan_id,
+                        $notifiable->name,
+                        $remover->name,
                     ),
+                ],
+                [
+                    'name' => 'Reason',
+                    'value' => $this->removalReason,
                 ],
                 [
                     'name' => sprintf(
                         '%s / %s',
-                        $this->member->division->locality('platoon'),
-                        $this->member->division->locality('squad')
+                        $notifiable->locality('platoon'),
+                        $notifiable->locality('squad')
                     ),
-                    'value' => $this->member->squad ? sprintf(
+                    'value' => $this->squad ? sprintf(
                         '%s / %s',
-                        $this->member->platoon->name,
-                        $this->member->squad->name
+                        $this->squad->platoon->name,
+                        $this->squad->name,
                     ) : 'Unassigned',
                 ],
                 [
                     'name' => $handle->label ?? 'In-Game Handle',
                     'value' => $handle->pivot->value ?? 'N/A',
                 ],
-            ])
-            ->info()
+            ])->error()
             ->send();
     }
 }
