+++ app/Notifications/Channel/NotifyDivisionMemberRequestApproved.php
@@ -1,39 +1,41 @@
 <?php
 
-namespace App\Notifications;
+namespace App\Notifications\Channel;
 
 use App\Channels\BotChannel;
 use App\Channels\Messages\BotChannelMessage;
 use App\Models\Member;
-use App\Models\MemberRequest;
 use App\Models\User;
+use App\Traits\DivisionSettableNotification;
+use App\Traits\RetryableNotification;
+use Exception;
 use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Notification;
 
-class MemberRequestPutOnHold extends Notification implements ShouldQueue
+class NotifyDivisionMemberRequestApproved extends Notification implements ShouldQueue
 {
-    use Queueable;
+    use DivisionSettableNotification, Queueable, RetryableNotification;
 
-    private MemberRequest $request;
+    private User $approver;
 
     private Member $member;
 
-    private User $approver;
+    private string $alertSetting = 'chat_alerts.member_approved';
 
     /**
      * Create a new notification instance.
      */
-    public function __construct(MemberRequest $memberRequest, User $approver, Member $member)
+    public function __construct(User $approver, Member $member)
     {
-        $this->request = $memberRequest;
-        $this->approver = $approver;
         $this->member = $member;
+        $this->approver = $approver;
     }
 
     /**
      * Get the notification's delivery channels.
      *
+     * @param  mixed  $notifiable
      * @return array
      */
     public function via()
@@ -51,8 +53,9 @@ class MemberRequestPutOnHold extends Notification implements ShouldQueue
     {
         return (new BotChannelMessage($notifiable))
             ->title($notifiable->name . ' Division')
+            ->target($notifiable->settings()->get($this->alertSetting))
             ->thumbnail($notifiable->getLogoPath())
-            ->message(addslashes("**MEMBER STATUS REQUEST ON HOLD** - :hourglass: A member status request for `{$this->member->name}` was put on hold by {$this->approver->name} for the following reason: `{$this->request->notes}`"))
+            ->message(addslashes("**MEMBER STATUS REQUEST** - :thumbsup: A member status request for `{$this->member->name}` was approved by {$this->approver->name}!"))
             ->success()
             ->send();
     }
