+++ app/Notifications/Channel/NotifyDivisionNewApplication.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Notifications;
+namespace App\Notifications\Channel;
 
 use App\Channels\BotChannel;
 use App\Channels\Messages\BotChannelMessage;
@@ -9,7 +9,7 @@ use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Notifications\Notification;
 
-class NewDivisionApplication extends Notification implements ShouldQueue
+class NotifyDivisionNewApplication extends Notification implements ShouldQueue
 {
     use Queueable, RetryableNotification;
